+++ src/player.hpp
@@ -24,7 +24,7 @@ class Player {
     void color(col tc);
     playertype type() const;
     void type(playertype t);
-    int find_best_move_index(int &move_no, float &score, const int max_search_level, bool use_pruning, bool search_until_no_captures);
+    int find_best_move_index(uint8_t &move_no, float &score, const int max_search_level, bool use_pruning, bool search_until_no_captures);
 };
 }
 #endif
